requirements.txt
argparse==1.3.0
certifi==14.05.14
dateutils==0.6.6
docutils==0.12
greenlet==0.4.5
motor==0.3.4
numpy==1.6.2
pymongo==2.7.1
python-dateutil==2.4.0
pytz==2014.10
scipy==0.15.1
six==1.9.0
tornado==4.0.2
wsgiref==0.1.2
